+++ ckb-debugger-tests/Cargo.toml
@@ -21,4 +21,3 @@ lazy_static = "1.4.0"
 log = "0.4.0"
 serde_json = "1.0"
 serde_plain = "1.0"
-sparse-merkle-tree = {version = "0.6.1"}
